+++ src/main/java/net/chimhaha/clone/web/dto/posts/PostsFindByCategoryResponseDto.java
@@ -1,4 +1,4 @@
-package net.chimhaha.clone.web.dto;
+package net.chimhaha.clone.web.dto.posts;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -9,19 +9,16 @@ import java.time.LocalDateTime;
 
 @Getter
 @NoArgsConstructor
-public class PostsFindByIdResponseDto {
+public class PostsFindByCategoryResponseDto {
     private Long id;
     private String title;
-    private String content;
     private String category;
     private LocalDateTime createdDate;
     private Integer views;
 
-    @Builder
-    public PostsFindByIdResponseDto(Posts posts) {
+    public PostsFindByCategoryResponseDto(Posts posts) {
         this.id = posts.getId();
         this.title = posts.getTitle();
-        this.content = posts.getContent();
         this.category = posts.getCategory();
         this.createdDate = posts.getCreatedDate();
         this.views = posts.getViews();
