+++ src/test/java/net/chimhaha/clone/service/PostsServiceTest.java
@@ -3,9 +3,9 @@ package net.chimhaha.clone.service;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import net.chimhaha.clone.domain.posts.Posts;
 import net.chimhaha.clone.domain.posts.PostsRepository;
-import net.chimhaha.clone.web.dto.PostsFindByCategoryResponseDto;
-import net.chimhaha.clone.web.dto.PostsFindByIdResponseDto;
-import net.chimhaha.clone.web.dto.PostsSaveRequestDto;
+import net.chimhaha.clone.web.dto.posts.PostsFindByCategoryResponseDto;
+import net.chimhaha.clone.web.dto.posts.PostsFindByIdResponseDto;
+import net.chimhaha.clone.web.dto.posts.PostsSaveRequestDto;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
